// src/app/models/register.model.ts
export interface register_interface{
    identificador: string;
    contraseña: string;
    usuario_Nombre: string;
    usuario_Apellidos: string;
    usuario_Edad: number;
    usuario_Tarjeta_Titular: string;
    usuario_Tarjeta_Direccion: string;
    usuario_Tarjeta_Numero_Tarjeta: string;
    usuario_Tarjeta_Fecha_Vencimiento: string;
}